requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI & LLM
google-generativeai==0.3.2
openai==1.10.0
anthropic==0.18.1

# Vector Database
pinecone-client==3.0.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
redis==5.0.1

# HTTP & API
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Data Processing
numpy==1.26.3
pandas==2.1.4
python-dotenv==1.0.0

# Image & Audio Processing
pillow==10.2.0
pydub==0.25.1

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.2.3

# Logging & Monitoring
loguru==0.7.2

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3twilio==9.10.1
